requirements.txt
docxtpl==0.3.1
Jinja2==2.9.4
lxml==3.7.2
MarkupSafe==0.23
python-docx==0.8.6
PyYAML==3.12
six==1.10.0
tqdm==4.11.0
